Setup dotfile storage consistently fixes #10
Previously the storage struct had to be setup with
multiple calls in different places:

 - In commands.go, Create a storage literal and set its fields
 - In files.go, call setup()
 - In data.go, call setPath()

These were called lots of different places in the code,
and there was no guarantee that the struct had to been
initialized correctly. Instead I created one function, Setup(),
to do this work. The fields the CLI must call this now.
Once Setup() is called, the backend is free to use the storage
struct.