Make initial commit on init and notify user if nothing happened
Previously `init` did not make an initial commit. This would break
operations like log or checkout.

Another problem was that `commit` and `init` did not tell the user
when nothing happened. Init will now tell you what file the
alias is tracking if you attempt to initialize it again.

This is helpful if you want to have different aliases/commit histories
for the same file.


Former-commit-id: e4c497cdefded090ccdb4e75e23b93016796f462