Add cmd directory and add go modules for dot and dotfilehub
The cmd directory holds go modules to create the binaries. The
directories within it are named the same as the exported binary.

This is inspired from
https://github.com/golang-standards/project-layout/tree/master/cmd

I didn't choose to add the pkg or internal directories reccomended
because I don't see the point yet. Maybe down the road.

Update Makefile accordingly.