Create page struct to store session data
I filled out the session database code to get the id and username of the
logged in user. This gets loaded at the start of http handlers and
stored in the page struct. type Page serves both as a place to bundle all the
relevant info about the page to be rendered, and as the struct that
templates read.

Change the login link to be the username when the session is found.