Rename table: temps => temp_files
Previously temp was setup to link to an already created file. This
wouldn't work when creating a file.

My intended flow is for a user to submit the file they want to create
in a form, and then the next page will show them their formatted temp
file, and give them a choice to either create it or discard.

This makes implementing the file.Storer requirements straightforward.

MIGRATE: drop table temps;