Add transaction field to Storage
Basically filling in the rest of the logic. The
storage struct should be mostly ready to go. This is all untested
so it will likely be changed. The new idea in this
commit is adding a transaction field to Storage.

The idea is allow for easy roll backs since some operations are doing many
different executions. Init is one good example of this.