Refactor database storage initialization
I created a struct called stagedFile which handles the logic of
combining files and temp files. Now storage only has to load a staged
file and the rest of the interface should work. If there is a temp
file, it will put its contents in DirtyContent, meaning its an unsaved
change to the file. Otherwise it will just use the current state of
the file.

Storage is pretty well tested and should be ready to start integrating
with the web server.