Fix temp file flow for new files / file edits
Previously the query param for the edit button didn't work. I fixed
this with a hidden input and changing the method to GET. It seems kind
of hacky but the example was in the MDN.

I fixed logic in handlers for choosing when to load the previous temp
file.

I changed input types in index.tmpl to "search". I learned about this
today while reading MDN as well. Such a great resource!