Add diffmatchpatch as dependency
Previously I was using the OS's diff binary. This worked fine for the
CLI. However the web server also needs to make diffs. I added a
file.Diff function in storer that returns []diffmatchpatch.Diff which
either the server or CLI can use to display.

Unfortunately the library doesn't provide an option that is close to
the -u version on diff. It includes all the unchanged text which isn't
interesting. So this has to be processed again so that it only shows a
few lines of unchanged text for context. I still need to implement
this for the web server, which is why there is a pointless looking for
loop in this commit.