Create a base template
I changed the method of rendering templates. This declares a single
base template which is always rendered. The base template calls a
`content` function that is created on render. The `content` function
executes a specific page template. This is nice because it allows for
all the templates to be parsed and loaded in at startup.

This system was inspired by:
https://www.kylehq.com/2017/05/golang-templates---what-i-missed/

Previously template included the nav, and head. This was
incorrect HTML because there was no <html> tag. I didn't want to
have to include that and close it on every page. It led to frequent
bugs where I would forget to include a template, or do it in the wrong
order.

This introduces a new template directory structure: base.tmpl,
`user/`, `file/`, and `static/`.