Fix issue where "restore" button displayed in file view
This template handles both the file view and commit view. The
variables are a bit confusing in here - .Data.hash refers to the file
or commits hash, and .Vars.hash refers to the hash variable in the
URL. The template uses .Vars.hash to decide whether it is a commit
view or not.