Fix tests after commit timestamp change
I recently changed the commit timestamp column to be unix time rather
than a datetime. This broke the initTestFileAndCommit function
because it expects an order of newest to oldest for the commit
list. The initial and new commit were both being created on the same
second. This is the same timestamp in unix time.

I fixed this by adding a time.Sleep(time.Second).