Implement pull and print CLI commands
This comes with various changes to local storage and the storer
interface. Most notably file.Storer is now an io.Closer, and file
operations that modify content will close it after finishing. In the
local package this means writing the JSON file, and in the db it means
closing the transaction. This allows methods to be called multiple
times before the file schemas is updated.

I also split up the local storage initialization between two methods:
one for creating the struct, and one for setting the file. This allows
it to be usable without a current file that is tracked. It will also
allow it to be reused for multiple files, E.G pull all.

I added the print command to help debug the pulls, but I'm going to
leave it because it seems useful for general use. It prints the
current content of the file or its json schema with --json.