Refactor local.TrackedFile => file.TrackingData
Because the server API needs to use this structure, it makes more
sense for this struct to live in the file package. I think the new
name is more descriptive as well.

This also allows for storer.SaveCommit to require a Commit struct
instead of seperate args for hash, message, and timestamp. This will
ensure that commit timestamps are always saved in unix time.