Move local push and pull to be pointer receivers
This simplifies calling these methods. Previously they had
arguments the file alias which was already stored in storage. I also
had to add user config to storage so that they can access those
settings. This will be helpful in the future for all storage methods
if additional user preferences are added.