Refactor config setup; fix misc bugs found from test failures
NewStorage is passed the config path. I changed the fallback
config path to ~/.dotfile-config.json.

I renamed LoadFile to SetTrackingData and also moved the definition of
Close so the diff looks funky around there.

I refactored local.InitFile to a be pointer receiver which makes its
definition simpler. Some of the extra work is now done in the CLI
package. This makes it more consistent with the other CLI
functions.