Factor out the content column from the files table
The files table now stores a reference to its current
commit. Initially I was hesitant to do this because it makes the
relationships more complex and the performance hit
for getting a files content. This is because content always needs to
be decompressed now. Previously it was stored in plain text in
the file.

This comes with many supporting changes. I split db.Storage into
db.FileTransaction and db.FileContent. This is possible from the
file interface changes in the previous content. This makes the setup
for file operations easier to follow.

This commit comes with a few changes for the API.
Push is still a WIP. It will be more straightforward to
implement with this change.