Implement timezone setting
This allows a user to set their timezone preference. It's not obvious
how to do it yet, but it expects a tzdata style string like
America/Los_Angeles

I also made all the timestamp loading more consistent throughout the
db package.

Finally, I updated the layout in the settings template.