Implement pull --all and simplify local storage
Storage methods that contact the API now take a dotfileclient as an
argument. This makes them more flexible. This allows the UserConfig
field to be removed from Storage - it was only used for a few methods
and made initialization more complex. Now the user config is only read
when it needs to be.