Remove state in local storage and fix bugs in pull
Pull was hard to reason about because it wasn't clear when
storage was tracking a file or not. I removed the constructor for
Storage and instead have callers use a literal. Passing home around
everywhere saved a few function calls but it made the code more
cumbersome.

This fixes pull --create-dirs.