Add forked from username to commit summary
Previously the forked from URL was stored in the commit message.
If the forked from ID was not null the commit message was marked up as
a link. This broke in the case that the original file was
deleted - it left the specialized commit message but forked from was
null.

This resolves that issue by adding an additional query to determine
the forked from username.