Change db functions to take a connection as an argument
The executor interface leys database functions use either a
transaction or a normal connection. Previously transactions were only
used in some functions. Sometimes the caller was expected to close
them, and sometimes not.

My new strategy is to always create, rollback, and commit transactions
in the same funciton. This removes ambiguity and should prevent
database locks.