Fix diff - strip carriage returns from web content
I noticed that diffs on the web had a lots of extra newlines. After
investigating I found that there were carraige return characters being
added. I don't really understand why still. Maybe it's just a side
effect of <textarea> elements? It only happens on content added from
the web.

I also cleaned up the diff code a bit and removed the ShortenEqualText
function which wasn't doing anything. I added the DiffPrettyText/HTML
functions so to the dotfile package so its easy to see them both in
one place.

The DiffPrettyText function is almost identical to the diffmatchpatch
package version. However the dmp function is a pointer reciever so its
not convenient to use.