Add flag to CLI for changing the config file path and various QOL
I pushed a change last night that caused the Travis CI build to fail
even though the tests were passing locally. I realized that this was
because the cli tests were using my actual config file, and the CI
server didn't have this setup.

To resolve this I added a path argument to local.ReadConfig and
local.SetConfig. The failing test also made me realize that the
dotfileclient doesn't always need to be setup for list and show, so I
refactored them. I also fixed an outstanding bug in show where pulling the
remote json data wasn't indented.

I added a feature to the config command that prints the current value
when value is absent. For example `dotfile config remote` now prints
https://dotfilehub.com instead of setting remote to an empty string.