Fix timezones - revamp user sessions
Previously the timezone was pulled as part of a query. This meant that
the timezone would always be that files owner - not the currently
logged in user. To fix this I initially thought to just add a timezone
field to the Session record, but then I realized that the session
record was basically a duplication of the user record. And plus, it
wasn't a record anymore, it was more than that. Instead I created the
UserSession struct, which models the session and users tables joined
together.

I had to refactor a bunch with the UserSession struct. I also made
getter methods for the page to access these fields to avoid any
accidently nil pointer dereferences - previously this was just assumed
to not be possible because of protected handlers, but this didn't seem
very safe.

As part of the refactoring I removed the db.User function. I figured
that only an authenticated session should have access to user
data, so I made it only accessible with the db.Session function. The
API handlers was using this in a few places, but I factored it out by
using userID in more places.