Downgrade to go-org@v1.3.2
Newer versions of go-org rewrite links with a .org extension to .html.
This breaks links in README.org. README.org uses relative links so
that it works within Github or within a Dotfilehub server. Because of
this, it's necessary for the links in the exported html to have .org
extensions.

https://github.com/niklasfasching/go-org/commit/84d56e95624f3ae1099edb9b527d4eb4a5df4e1d